refactor: integrate app logic into execution module
